+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - PR #3608 Update OPS codeowner group name
 - PR #3431 Port NVStrings translate to cudf strings column
 - PR #3587 Merge CHECK_STREAM & CUDA_CHECK_LAST to CHECK_CUDA
+- PR #3612 Add ability to customize the JIT kernel cache path
 
 ## Bug Fixes